src/server_py/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart
pydantic==2.5.2
google-generativeai
requests
pillow
